requirements.txt
torch
torchvision
matplotlib
numpy
pillow
ultralytics
flask
werkzeug
opencv-python